requirements.txt
# Google API client libraries for Tasks integration
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# HTTP library for webhook posting
requests>=2.28.0
